requirements.txt
flask
flask-cors
google-generativeai
huggingface_hub
python-dotenv
gunicorn
